refactor: consolidate navigation items and not-found components
